// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/recipe-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecipeImageHttpResponse } from './recipe-image-http-response';
import { RecipeStepHttpResponse } from './recipe-step-http-response';

export interface RecipeHttpResponse {
    category?: string | null;
    comment?: string | null;
    cookTime?: number | null;
    createdAt?: string;
    description?: string | null;
    favoriteRecipeId?: string | null;
    id?: string;
    imageAssetId?: string | null;
    imageUrl?: string | null;
    images?: Array<RecipeImageHttpResponse>;
    isFavorite?: boolean;
    isNutritionAutoCalculated?: boolean;
    isOwnedByCurrentUser?: boolean;
    language?: string;
    languageConfirmed?: boolean;
    manualAlcohol?: number | null;
    manualCalories?: number | null;
    manualCarbs?: number | null;
    manualFats?: number | null;
    manualFiber?: number | null;
    manualProteins?: number | null;
    missingIngredientCount?: number;
    name?: string;
    prepTime?: number | null;
    qualityGrade?: string;
    qualityScore?: number;
    servings?: number;
    steps?: Array<RecipeStepHttpResponse>;
    totalAlcohol?: number | null;
    totalCalories?: number | null;
    totalCarbs?: number | null;
    totalFats?: number | null;
    totalFiber?: number | null;
    totalProteins?: number | null;
    usageCount?: number;
    visibility?: string;
}